+++ src/routes/+layout.svelte
@@ -3,31 +3,19 @@
 
 	import Menubar from '$lib/menu/Menubar.svelte';
 	import Contextbar from '$lib/menu/Contextbar.svelte';
+	import Statusbar from '$lib/menu/Statusbar.svelte';
+	import Toasts from '$lib/menu/Toasts.svelte';
 
-	import { Button } from 'bits-ui';
-	import Icon from '@iconify/svelte';
-
-	import { expoOut } from 'svelte/easing';
-	import { fade, slide } from 'svelte/transition';
-	import { onDestroy, onMount } from 'svelte';
-	import { listen, type UnlistenFn } from '@tauri-apps/api/event';
-	import { writeText } from '@tauri-apps/plugin-clipboard-manager';
+	import { onMount } from 'svelte';
 	import NavbarLink from '$lib/menu/NavbarLink.svelte';
 	import InstallProgressPopup from '$lib/modlist/InstallProgressPopup.svelte';
 	import WelcomePopup from '$lib/menu/WelcomePopup.svelte';
-	import { clearToast, toasts } from '$lib/toast';
 	import { refreshColor } from '$lib/theme';
-
-	let status: string | null = null;
-	let unlisten: UnlistenFn | undefined;
+	import InstallModPopup from '$lib/modlist/InstallModPopup.svelte';
 
 	onMount(async () => {
 		refreshColor('accent');
 		refreshColor('primary');
-
-		unlisten = await listen<string | null>('status_update', (evt) => {
-			status = evt.payload;
-		});
 	});
 </script>
 
@@ -55,59 +43,10 @@
 		<slot />
 	</div>
 
-	{#if status !== null}
-		<div
-			class="border-primary-600 text-primary-400 flex w-full items-center border-t px-3 py-1 text-sm"
-			transition:slide={{ duration: 200, easing: expoOut }}
-		>
-			<Icon icon="mdi:loading" class="animate-spin" />
-			<span class="ml-2">{status}</span>
-		</div>
-	{/if}
-
-	<div
-		class="absolute right-0 bottom-0 z-10 flex max-w-[50rem] flex-col items-end justify-end gap-1 p-2 xl:max-w-[90rem]"
-	>
-		{#each $toasts as toast, i}
-			<div
-				class="flex items-start overflow-hidden rounded-md p-1.5 xl:p-2 xl:text-lg {toast.type ===
-				'error'
-					? 'bg-red-600'
-					: 'bg-accent-600'}"
-				in:slide={{ duration: 150, easing: expoOut }}
-				out:fade={{ duration: 100 }}
-			>
-				<div class="mt-auto mr-3 grow overflow-hidden px-2">
-					{#if toast.name !== undefined}
-						<span class={toast.type === 'error' ? 'text-red-200' : 'text-accent-200'}
-							>{toast.name} -</span
-						>
-					{/if}
-
-					<span class="font-medium break-words text-white">{toast.message}</span>
-				</div>
-
-				{#if toast.type === 'error'}
-					<Button.Root
-						class="rounded-xs p-1 hover:bg-red-500"
-						on:click={() => writeText('`' + toast.name + ' - ' + toast.message + '`')}
-					>
-						<Icon icon="mdi:clipboard-text" class="text-primary-100 text-lg" />
-					</Button.Root>
-				{/if}
-
-				<Button.Root
-					class="rounded-md p-1 {toast.type === 'error'
-						? 'hover:bg-red-500'
-						: 'hover:bg-accent-500'}"
-					on:click={() => clearToast(i)}
-				>
-					<Icon icon="mdi:close" class="text-primary-100 text-lg" />
-				</Button.Root>
-			</div>
-		{/each}
-	</div>
+	<Statusbar />
+	<Toasts />
 </main>
 
+<InstallModPopup />
 <InstallProgressPopup />
 <WelcomePopup />